+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 0.2.0 (Next)
 
-* Your contribution here.
+* Replaces non-breaking spaces (`&nbsp;`) with regular spaces in converted Markdown - [@dblock](https://github.com/dblock).
 
 ### 0.1.0 (2026/09/19)